[useScrollLock] Restore shared touch helpers
